requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
mysql==0.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
protobuf==3.19.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise
